rename plugin_util to plugin_loader, fix format, add comment
