+++ source/adaptive_load/plugin_loader.cc
