+++ test/adaptive_load/BUILD
@@ -39,11 +39,11 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "plugin_util_test",
-    srcs = ["plugin_util_test.cc"],
+    name = "plugin_loader_test",
+    srcs = ["plugin_loader_test.cc"],
     repository = "@envoy",
     deps = [
-        "//source/adaptive_load:plugin_util",
+        "//source/adaptive_load:plugin_loader",
         "@envoy//source/common/protobuf:utility_lib_with_external_headers",
         "@envoy//test/test_common:utility_lib",
     ],
